Version 1.2

+++ clean_cluster1.sh
@@ -11,5 +11,9 @@ kubectl delete -f cluster1.yaml
 kubectl delete -f backup_cluster1.yaml
 
 # AWS delete
-aws s3 rm --recursive ${s3_cluster1}
-
+if [ "${OBJECT_STORAGE}" == "AWS" ]; then
+  aws s3 rm --recursive ${s3_cluster1}
+else
+  # Stop MinIO
+  docker ps | grep minio | awk '{print $1}' | xargs -I % docker stop %
+fi
